requirements.txt
Django==2.1.7
djangorestframework==3.9.2
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.9
